crucial fixes
fixed some bugs that actually broke the entire network (mainly activating and feeding forward in the wrong order, leading to 0 output. Also fixed a bug that resulted in unconnected nodes 1 gen after loading a file.
Game changes: after a goal, the ball gets a small, random velocity to increase general learning (otherwise, if one player out of two makes a goal during the time limit, he will again and again and again, due to the determinsitic nature of the game and network)